cpp(7-3): 횟수를 반환하는 함수의 반환값을 float에서 int로 변경

+++ cppPractice/practice7/7-3.cpp
@@ -29,7 +29,7 @@ public:
 		float fl_count = count;
 		return fl_total / fl_count;
 	}
-	float get_count() { return count; }
+	int get_count() { return count; }
 };
 
 int main() {
